* 'master' of ssh://oej@git.sip-router.org/sip-router:
  makefile: even more quiet
  modules warning fixes
  * callcontrol
  makefile: quiet output support
  pike_top: moved from tools/ to utils/
  kamctl: moved from tools/ to utils/
  core ev: minor cleanups
  topoh: new module for hiding topology details
  core: new sr events system
  core: execute callbacks for NET_DATA_IN and NET_DATA_OUT

+++ lib/binrpc/binrpc_api.c
@@ -457,7 +457,7 @@ static int get_reply(struct binrpc_handle *handle,
 			if (n == 0)
 				snprintf(binrpc_last_errs, sizeof(binrpc_last_errs)-1,
 					"get_reply: read unexpected EOF: received %d bytes"
-					" of reply", crt - handle->buf);
+					" of reply", (int)(long)(crt - handle->buf));
 			else
 				snprintf(binrpc_last_errs, sizeof(binrpc_last_errs)-1,
 					"get_reply: read reply failed: %s (%d)",
